.github/workflows/scripts/requirements.txt
PyGithub==1.55
dpath==2.0.1
GitPython==3.1.18
ruamel.yaml==0.17.10
PyPAC==0.15.0